Update data_model.yaml
Added note about Need in Child table as I don't feel it's adequate for describing children as children can have many needs. This field could refer to more generalized situations, but there seems to be a need for more details on this aspect, so both structures have tentatively been left in place.